+++ src/constants/emoji.constant.js
@@ -1,0 +1,14 @@
+const EMOJI = {
+  thumbs_up: '👍',
+  heart: '❤️',
+  clap: '👏',
+  lion: '🦁',
+  thinking: '🤔',
+  smile: '☺️',
+  clover: '🍀',
+  eyes: '👀',
+  perfect: '💯',
+  bulb: '💡',
+};
+
+export default EMOJI;
